Fixed an error occured on CCL in present-object displaying a call with UnknownArguments.
This error was a CCL specific, I encounter it using Rove unittest, which uses
dissect framework to print tracebacks of catched signals. When you call
a function with wrong number of arguments, CCL generates a stacktrace where some
call has an object of type DISSECT:UNKNOWN-ARGUMENTS, and present-object fails to
print the arguments for such call, because think that arguments should be a list.